+++ Changelog.md
@@ -2,7 +2,8 @@
 
 ## v1.3.0 (in development)
 
-- [WIP] Added a way to define the element when defining a bemto block.
+- Added a way to define the element when defining a bemto block.
+- Added a way to pass content to those elements from the block call.
 - Added a way to pass options object as well/instead of a tagString.
 - Made __BemtoElem API a bit easier to use.